fix: change datatype to integer for id

+++ lostandfound/src/main/java/com/example/lostandfound/repository/MatchRepository.java
@@ -1,10 +1,11 @@
 package com.example.lostandfound.repository;
 
-import com.example.lostandfound.model.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.lostandfound.model.Match;
+
 @Repository
-public interface MatchRepository extends JpaRepository<Match, Long> {
+public interface MatchRepository extends JpaRepository<Match, Integer> {
     // Additional custom queries can be added here
 }
