+++ lostandfound/src/main/java/com/example/lostandfound/repository/UserRepository.java
@@ -1,11 +1,12 @@
 package com.example.lostandfound.repository;
 
-import com.example.lostandfound.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.lostandfound.model.User;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
     // Additional custom queries can be added here
 }
